refactored entry scripts:
1. execution script renamed to 'run_oktoberfest.py' to prevent import problems, as 'oktoberfest.py' would superseed the package namespace oktoberfest
2. fixed imports in oktoberfest.py -> runner.py, which are now relative since this file is now part of the oktoberfest package when imported
3. import runner.py in __init__.py and moved version and copyright info to this file
4. created new run_oktoberfest.py which is now properly importing the oktoberfest package, printing version and copyright and executing the runner.
